+++ inst/shiny/schedule_builds/app.R
@@ -5,7 +5,6 @@ library(googleCloudRunner)
 ui <- fluidPage(
   shinyjs::useShinyjs(),
   titlePanel("Cloud Build Scheduler"),
-
   sidebarLayout(
     sidebarPanel(
       helpText("Upload a cloudbuild.yml file to schedule"),
@@ -16,15 +15,27 @@ ui <- fluidPage(
                   "application/x-yaml",
                   "text/vnd.yaml"
                 )),
-      textInput("schedule_name", label = "Schedule name", value = "my-schedule"),
+      textInput("schedule_name", label = "Schedule name",
+                value = paste0("my-schedule-", format(Sys.time(), "%Y%m%d%H%M%S"))),
       textInput("cron", "cron schedule", value = "15 9 * * *"),
+      h4("Project:", cr_project_get()),
+      p(a(href=sprintf("https://console.cloud.google.com/cloudscheduler?project=%s",
+                       cr_project_get()), "Project schedule listings")),
+      p(a(href=sprintf("https://console.cloud.google.com/cloud-build/builds?project=%s",
+              cr_project_get()), "Project build history")),
       uiOutput("logs_link")
   ),
     mainPanel(
-      helpText("This app will take an existing cloudbuild.yml, validate a build and then schedule"),
+      helpText("This app will take an existing cloudbuild.yml, validate a build and then schedule it"),
+      helpText("Create the cloudbuild.yml file either via",
+               a(href="https://code.markedmondson.me/googleCloudRunner/reference/cr_build_write.html",
+                 "googleCloudRunner's cr_build_write() function"),
+               "or manually following the",
+               a(href = "https://cloud.google.com/cloud-build/docs/build-config",
+                 "Cloud Build yaml schema")),
       shinyjs::hidden(h3(id = "working", "Building...")),
       p(textOutput("validated")),
-      textOutput("scheduled")
+      uiOutput("scheduled")
   )
 )
 )
@@ -40,7 +51,7 @@ server <- function(input, output, session){
   output$logs_link <- renderUI({
     req(do_build())
 
-    a(href=bb$metadata$build$logUrl, "Build Log")
+    a(href=do_build()$metadata$build$logUrl, "Build Log")
 
   })
 
@@ -61,7 +72,7 @@ server <- function(input, output, session){
     "Build error - check logs"
   })
 
-  output$scheduled <- renderText({
+  output$scheduled <- renderUI({
     req(built())
 
     if(built()$status == "SUCCESS"){
@@ -72,20 +83,23 @@ server <- function(input, output, session){
                   description = "Scheduled via googleCloudRunner shiny",
                   overwrite = TRUE)
 
+      if(!is.null(x$state))
       return(
-        paste(
-        "==CloudScheduleJob==\n",
-        "name: ", x$name,
-        "state: ", x$state,
-        "httpTarget.uri: ", x$httpTarget$uri,
-        "httpTarget.httpMethod: ", x$httpTarget$httpMethod,
-        "userUpdateTime: ", x$userUpdateTime,
-        "schedule: ", x$schedule,
-        "timezone: ", x$timeZone)
+        tagList(
+          h3("Cloud Build Scheduled"),
+          tags$ul(
+            tags$li(paste("name: ", x$name)),
+            tags$li(paste("state: ", x$state)),
+            tags$li(paste("userUpdateTime: ", x$userUpdateTime)),
+            tags$li(paste("schedule: ", x$schedule)),
+            tags$li(paste("timezone: ", x$timeZone)),
+          )
+
+        )
       )
     }
 
-    "Not scheduling"
+    h3("Build Error - not scheduling - see build logs")
 
     })
 }